Introduce controller tests, implement validation service, and perform refactoring of utility files
