+++ src/server/api/services/ParserService.js
@@ -1,4 +1,4 @@
-const { URL_REGEX, API_DEFINITION_REGEX, SUB_DOMAIN_REGEX, KEYWORD_REGEX } = require('../constants/Constants');
+const { URL_REGEX, API_DEFINITION_REGEX } = require('../constants/Constants');
 
 module.exports = {
   /**
@@ -8,7 +8,7 @@ module.exports = {
    * @param {string} text - The input text to parse and extract URLs from.
    * @throws {Error} If there is an error during the parsing process, it will be thrown.
  */
-  parsing: function (text) {
+  parsingService: function (text) {
     try {
       const matchUrls = text.match(URL_REGEX);
       matchUrls.forEach((url) => {
