+++ src/server/config/routes.js
@@ -11,6 +11,6 @@
 module.exports.routes = {
 
   'POST /api/v1/run/crawler': 'CrawlingController.startCrawling',
-  'GET /api/v1/download/process/index-files': 'DownloadAndProcessIndexFilesController.startDownloadAndProcessIndexFiles'
+  'GET /api/v1/process/index-files': 'DownloadAndProcessIndexFilesController.startDownloadAndProcessIndexFiles'
 
 };
