+++ src/server/package.json
@@ -31,6 +31,8 @@
     "sails-hook-orm": "^4.0.0",
     "sails-hook-sockets": "^2.0.0",
     "sails-mongo": "^2.0.1",
+    "swagger-parser": "^10.0.3",
+    "yaml": "^2.3.1",
     "zlib": "^1.0.5"
   },
   "devDependencies": {
